ARM: imx: busfreq: Only adjust cpu frequency if the driver is available
The busfreq driver is relying on the cpu frequency scaling driver for
some structures and functionality.  If you build the platform without
cpufrequency scaling enabled then these symbols are undefined.  Therefore
we only include these codepaths if the driver is built in.

It may make more sense to turn the busfreq driver into a Kconfig option
that depends on the cpu frequency scaling.

Signed-off-by: Jon Nettleton <jon@solid-run.com>