MXC video: Increase the vertical blanking intervall for some VESA modes
The timing of several VESA modes is calculated using a special formula.
Unfortunately, the result may not be acceptable by the IPU because the
lower margin of the vertical blanking intervall is 1. This patch adds
a hack that increases both, lower_marging and vsync_len (which in this
case is also too low) by 2 in order to make these modes work.

Signed-off-by: Rudi <r.ihle@s-t.de>